Add confirmation to user deletion

+++ frontend/src/Manage/Admin/DeleteUser/DeleteUser.jsx
@@ -16,25 +16,27 @@ export const DeleteUser = ({ fetchUsers, users }) => {
 
   const handleDelete = async (event) => {
     event.preventDefault()
-    const id = users.filter(user => user.email === selectedUser)[0].id
-    try {
-      const resp = await fetch(`/api/users/${id}/`, {
-        method: 'DELETE',
-        headers: {
-          'Content-Type': 'application/json'
+    if (window.confirm(t('DELETE_USER_CONFIRM'))) {
+      const id = users.filter(user => user.email === selectedUser)[0].id
+      try {
+        const resp = await fetch(`/api/users/${id}/`, {
+          method: 'DELETE',
+          headers: {
+            'Content-Type': 'application/json'
+          }
+        })
+        if (resp.status === 200) {
+          setToastData({ message: t('USER_DELETE_SUCCESS'), type: 'success' })
+        } else if (resp.status === 401) {
+          redirectToPage(LOGIN_PAGE)
+        } else {
+          setToastData({ message: t('USER_DELETE_ERROR'), type: 'danger' })
         }
-      })
-      if (resp.status === 200) {
-        setToastData({ message: t('USER_DELETE_SUCCESS'), type: 'success' })
-      } else if (resp.status === 401) {
-        redirectToPage(LOGIN_PAGE)
-      } else {
-        setToastData({ message: t('USER_DELETE_ERROR'), type: 'danger' })
+      } catch (error) {
+        setToastData({ message: t('UNEXPECTED_ERROR'), type: 'danger' })
       }
-    } catch (error) {
-      setToastData({ message: t('UNEXPECTED_ERROR'), type: 'danger' })
+      fetchUsers()
     }
-    fetchUsers()
   }
 
   return (
